+++ gameplay/src/PhysicsCollisionObject.cpp
@@ -34,7 +34,7 @@ struct CollidesWithCallback : public btCollisionWorld::ContactResultCallback
     bool result;
 };
 
-PhysicsCollisionObject::PhysicsCollisionObject(Node* node, short group, short mask)
+PhysicsCollisionObject::PhysicsCollisionObject(Node* node, int group, int mask)
     : _node(node), _collisionShape(NULL), _enabled(true), _scriptListeners(NULL), _motionState(NULL), _group(group), _mask(mask)
 {
 }
